compile under jdk1.2.2
git-svn-id: https://svn.apache.org/repos/asf/jakarta/poi/trunk@352677 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/record/formula/AttrPtg.java
@@ -75,7 +75,7 @@ import java.util.List;
 public class AttrPtg
     extends OperationPtg
 {
-    public final static short sid  = 0x19;
+    public final static byte sid  = 0x19;
     private final static int  SIZE = 4;
     private byte              field_1_options;
     private short             field_2_data;
